+++ modules/micropython.cmake
@@ -1,5 +1,4 @@
 # Auto-generated - include all compiled modules
-include(${CMAKE_CURRENT_LIST_DIR}/usermod_lvgl/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_algorithms/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_async_demo/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_bitwise/micropython.cmake)
@@ -36,6 +35,5 @@ include(${CMAKE_CURRENT_LIST_DIR}/usermod_string_operations/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_super_calls/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_traits/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_tuple_operations/micropython.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}/usermod_typed_funcs/micropython.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/usermod_sensor_lib/micropython.cmake)
-include(${CMAKE_CURRENT_LIST_DIR}/usermod_lvui/micropython.cmake)
-include(${CMAKE_CURRENT_LIST_DIR}/usermod_lvgl_mvu/micropython.cmake)